YOOOO COMPLETED

+++ healthy_nutrition_bot/domain/entities/usdafood/FoodNutrientResponse.cs
@@ -3,9 +3,9 @@
 public class FoodNutrientResponse
 {
     public string FoodName { get; set; } = string.Empty;
-    public double Calories { get; set; }
-    public double Protein { get; set; }
-    public double Fat { get; set; }
-    public double Carbohydrates { get; set; }
+    public float Calories { get; set; }
+    public float Protein { get; set; }
+    public float Fat { get; set; }
+    public float Carbohydrates { get; set; }
     public bool IsHealthy { get; set; }
 }
